refactor: provide metadata loader with factory fn
